Setup roles and authorizations

+++ src/main/java/br/com/mundim/RestaurantReservationManagment/service/CustomerService.java
@@ -5,6 +5,7 @@ import br.com.mundim.RestaurantReservationManagment.model.dto.CustomerDTO;
 import br.com.mundim.RestaurantReservationManagment.model.entity.Customer;
 import br.com.mundim.RestaurantReservationManagment.repository.CustomerRepository;
 import br.com.mundim.RestaurantReservationManagment.security.AuthenticationService;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -35,7 +36,13 @@ public class CustomerService {
         return customerRepository.findAll();
     }
 
+    public Customer findCurrentCustomer() {
+        UserDetails user = authenticationService.findUserByBearer();
+        return customerRepository.findByEmail(user.getUsername());
+    }
+
     public Customer findById(Long customerId) {
+        authenticationService.verifyCustomerOwnership(customerId);
         return customerRepository.findById(customerId)
                 .orElseThrow(() -> new BadRequestException(
                         CUSTOMER_NOT_FOUND_BY_ID.params(customerId.toString()).getMessage())
@@ -46,6 +53,7 @@ public class CustomerService {
         Customer customer = customerRepository.findByCpf(cpf);
         if (customer == null)
             throw new BadRequestException(CUSTOMER_NOT_FOUND_BY_CPF.params(cpf).getMessage());
+        authenticationService.verifyCustomerOwnership(customer.getId());
         return customer;
     }
 
@@ -53,10 +61,12 @@ public class CustomerService {
         Customer customer = customerRepository.findByEmail(email);
         if (customer == null)
             throw new BadRequestException(CUSTOMER_NOT_FOUND_BY_EMAIL.params(email).getMessage());
+        authenticationService.verifyCustomerOwnership(customer.getId());
         return customer;
     }
 
     public Customer update(Long customerId, CustomerDTO dto) {
+        authenticationService.verifyCustomerOwnership(customerId);
         Customer customer = findById(customerId);
         customer.setName(dto.name());
         customer.setCpf(dto.cpf());
@@ -66,6 +76,7 @@ public class CustomerService {
     }
 
     public Customer deleteById(Long customerId) {
+        authenticationService.verifyCustomerOwnership(customerId);
         Customer customer = findById(customerId);
         customerRepository.deleteById(customerId);
         return customer;
